test(menu): migrate tests
