+++ core/src/components/menu/test/basic/menu.e2e.ts-snapshots/menu-basic-side-toggled-md-rtl-Mobile-Firefox-linux.png
